Modified README, function and comments added to main, new test for count_words()
Added section for word counter feature

The implementation of word_paragraph was changed, it now returns the actual list of words and not directly the length of it. The purpose of this is that the new function can take advantage of the already computed words and not do the work from scratch, along with the added bonus of making the implementation slightly cleanear.

As for test, it checks the type of the returned value of the new function, but more tests are needed.