Display 10 most frequent words in the file added
The word_count function had a small bug, it was only taking the first word and stopped counting, it is now fixed. Now it serves the functionality of dict_to_list() and show_words().

dict_to_list() takes the output word_count, creates a list of binary tuple which puts first the value and then the key of each dictionary entry. The purpose of this is having a way of organizing the list in a decending order.

As for show_words(), it simply iterates through the list of binary tuples and displays 10 most repeated words throughout the file.

All of the implement features and functions were tested in the test.py file, also they were implemented using the HtDF methodoloty (How to Define Functions).